+++ src/main/java/com/maxmind/geoip2/JsonInjector.java
@@ -3,6 +3,7 @@ package com.maxmind.geoip2;
 import com.fasterxml.jackson.databind.BeanProperty;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.InjectableValues;
+import com.maxmind.db.Network;
 import com.maxmind.geoip2.record.Traits;
 
 import java.util.List;
@@ -10,10 +11,12 @@ import java.util.List;
 class JsonInjector extends InjectableValues {
     private final List<String> locales;
     private final String ip;
+    private final Network network;
 
-    public JsonInjector(List<String> locales, String ip) {
+    public JsonInjector(List<String> locales, String ip, Network network) {
         this.locales = locales;
         this.ip = ip;
+        this.network = network;
     }
 
     @Override
@@ -25,8 +28,11 @@ class JsonInjector extends InjectableValues {
         if ("ip_address".equals(valueId)) {
             return ip;
         }
+        if ("network".equals(valueId)) {
+            return network;
+        }
         if ("traits".equals(valueId)) {
-            return new Traits(ip);
+            return new Traits(ip, network);
         }
         return null;
     }