+++ src/main/java/com/maxmind/geoip2/model/DomainResponse.java
@@ -2,6 +2,11 @@ package com.maxmind.geoip2.model;
 
 import com.fasterxml.jackson.annotation.JacksonInject;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
+import com.maxmind.db.Network;
+import com.maxmind.geoip2.NetworkDeserializer;
 
 /**
  * This class provides the GeoIP2 Domain model.
@@ -10,17 +15,27 @@ public class DomainResponse extends AbstractResponse {
 
     private final String domain;
     private final String ipAddress;
+    private final Network network;
 
     DomainResponse() {
         this(null, null);
     }
 
     public DomainResponse(
+            String domain,
+            String ipAddress
+    ) {
+        this(domain, ipAddress, null);
+    }
+
+    public DomainResponse(
             @JsonProperty("domain") String domain,
-            @JacksonInject("ip_address") @JsonProperty("ip_address") String ipAddress
+            @JacksonInject("ip_address") @JsonProperty("ip_address") String ipAddress,
+            @JacksonInject("network") @JsonProperty("network") @JsonDeserialize(using = NetworkDeserializer.class) Network network
     ) {
         this.domain = domain;
         this.ipAddress = ipAddress;
+        this.network = network;
     }
 
     /**
@@ -38,5 +53,16 @@ public class DomainResponse extends AbstractResponse {
     @JsonProperty("ip_address")
     public String getIpAddress() {
         return this.ipAddress;
+    }
+
+    /**
+     * @return The network associated with the record. In particular, this is
+     * the largest network where all of the fields besides IP address have the
+     * same value.
+     */
+    @JsonProperty
+    @JsonSerialize(using = ToStringSerializer.class)
+    public Network getNetwork() {
+        return this.network;
     }
 }